Moved plutins initializations after inner services initialized. closes #444

+++ src/xpiks-qt/Helpers/asynccoordinator.cpp
@@ -56,7 +56,7 @@ namespace Helpers {
 
     void AsyncCoordinator::reportStatus(AsyncCoordinator::CoordinationStatus status) {
         if (0 == m_StatusReported.fetchAndStoreOrdered(1)) {
-            emit statusReported(status);
+            emit statusReported((int)status);
         }
     }
 }
